Update README.md
Alex created this file and open it for the discussion. This file should be edited in the following way: 

1. create Github account
2. Fork this repository with the Readme.md file
3. Update the file in the way you want, commit it and push to your repository (only to your, not to OpenPTV)
4. Send us your pull request and we'll accept it if your file can be merged with the other editions and whether it does not contain mistakes and typos.

We recommend to try first http://try.github.io and also see the lectures on the site Software Carpentry, e.g. http://software-carpentry.org/4_0/vc/index.html and http://software-carpentry.org/blog/2012/06/git-tutorial-links.html